+++ src/main/java/com/gagechaeum/backend/security/service/CustomUserDetailsService.java
@@ -1,11 +1,10 @@
 package com.gagechaeum.backend.security.service;
 
 import com.gagechaeum.backend.security.account.domain.CustomUserDetails;
+import com.gagechaeum.backend.security.account.mapper.UserDetailsMapper;
 import com.gagechaeum.backend.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-
-import org.scoula.security.account.mapper.UserDetailsMapper;
 
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
