+++ src/main/java/com/gagechaeum/backend/user/service/UserServiceImpl.java
@@ -2,29 +2,40 @@ package com.gagechaeum.backend.user.service;
 
 import com.gagechaeum.backend.common.mail.MailService;
 import com.gagechaeum.backend.common.redis.RedisService;
-import com.gagechaeum.backend.global.exception.BusinessException;
-import com.gagechaeum.backend.global.exception.ErrorCode;
 import com.gagechaeum.backend.security.account.dto.UserLoginRequestDTO;
 import com.gagechaeum.backend.security.util.JwtUtil;
 import com.gagechaeum.backend.user.domain.User;
 import com.gagechaeum.backend.user.dto.TokenResponseDTO;
 import com.gagechaeum.backend.user.dto.UserJoinRequestDTO;
 import com.gagechaeum.backend.user.dto.UserResponseDTO;
+import com.gagechaeum.backend.user.exception.auth.*;
+import com.gagechaeum.backend.user.exception.signup.ValidationFailedException;
+import com.gagechaeum.backend.user.exception.verify.EmailAlreadyVerifiedException;
+import com.gagechaeum.backend.user.exception.verify.VerificationCodeExpiredException;
+import com.gagechaeum.backend.user.exception.verify.VerificationCodeMismatchException;
+import com.gagechaeum.backend.user.exception.verify.VerificationRateLimitException;
 import com.gagechaeum.backend.user.mapper.UserMapper;
+import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.DuplicateKeyException;
+import org.springframework.security.authentication.DisabledException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.RequestBody;
+
+import java.io.IOException;
 import java.time.LocalDateTime;
-import java.util.UUID;
+import java.util.*;
 
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserMapper userMapper;
+
     private final JwtUtil jwtUtil;
     private final RedisService redisService;
     private final PasswordEncoder encoder;
@@ -32,6 +43,7 @@ public class UserServiceImpl implements UserService {
 
     private final Logger log = LoggerFactory.getLogger(UserServiceImpl.class);
 
+    // 이메일 인증 정책
     private static final int VERIFY_MAX_TRIES = 5;
     private static final int DAILY_LIMIT = 10;
 
@@ -39,53 +51,71 @@ public class UserServiceImpl implements UserService {
         return userMapper.findByEmail(email) != null;
     }
 
+    // ===== 이메일 인증 요청 =====
     @Override
     public void requestEmailVerification(String email) {
+        // 이미 가입된 메일 검사
         if (userMapper.findByEmail(email) != null) {
-            throw new BusinessException(ErrorCode.USER_ALREADY_EXISTS);
+            throw new EmailAlreadyVerifiedException();
         }
+        // DB에서 이미 인증된 메일일 경우(기존 가입 이력) - 선택적으로 제한
         Boolean dbVerified = userMapper.selectIsVerifiedByEmail(email);
         if (Boolean.TRUE.equals(dbVerified)) {
-            throw new BusinessException(ErrorCode.EMAIL_ALREADY_VERIFIED);
+            // 필요시 별도 예외로 막아도 되고, 그냥 코드 발송 허용도 가능
+            // throw new EmailAlreadyVerifiedException();
         }
 
+        // 일일 전송 제한
         long dailyCnt = redisService.incrDailyAndSetExpireIfNew(email);
         if (dailyCnt > DAILY_LIMIT) {
-            throw new BusinessException(ErrorCode.VERIFICATION_RATE_LIMITED);
+            throw new VerificationRateLimitException();
         }
 
+        // 재전송 쿨타임
         if (redisService.hasEvCooldown(email)) {
-            throw new BusinessException(ErrorCode.VERIFICATION_RATE_LIMITED);
+            throw new VerificationRateLimitException();
         }
 
+        // 6자리 코드 생성
         String code = org.apache.commons.lang3.RandomStringUtils.randomNumeric(6);
+
+        // Redis 저장(코드/tries 초기화)
         redisService.saveEvCode(email, code);
+        // 쿨타임 60초
         redisService.setEvCooldown(email);
+
+        // 메일 발송
         mailService.sendVerificationCode(email, code);
     }
 
+    // ===== 인증 코드 확인 =====
     @Override
     public void confirmEmailVerification(String email, String code) {
         String saved = redisService.getEvCode(email);
         if (saved == null) {
-            throw new BusinessException(ErrorCode.VERIFICATION_CODE_EXPIRED);
+            throw new VerificationCodeExpiredException();
         }
 
         int tries = redisService.getEvTries(email);
         if (tries >= VERIFY_MAX_TRIES) {
-            throw new BusinessException(ErrorCode.VERIFICATION_RATE_LIMITED);
+            throw new VerificationRateLimitException();
         }
 
         if (!saved.equals(code)) {
             redisService.incEvTries(email);
-            throw new BusinessException(ErrorCode.VERIFICATION_CODE_MISMATCH);
+            throw new VerificationCodeMismatchException();
         }
 
+        // 성공: 가입 허용 플래그(30분) 세팅
         redisService.setEvOk(email);
+
         userMapper.updateIsVerifiedByEmail(email, true);
+
+        // 사용한 코드/시도정보 정리
         redisService.clearEvCodeState(email);
     }
 
+    // ===== 현재 인증여부 판단 (가입 전/후 공용) =====
     @Override
     public boolean isEmailVerifiedNow(String email) {
         if (redisService.hasEvOk(email)) return true;
@@ -93,78 +123,82 @@ public class UserServiceImpl implements UserService {
         return Boolean.TRUE.equals(v);
     }
 
+    // ===== 회원가입 시 인증검사 + is_verified=true 저장 =====
     @Override
     public UserResponseDTO registerUser(UserJoinRequestDTO req) {
-        log.info("🔒 회원가입 시도: {}", req.getEmail());
+        log.info("🔒 회원가입 시도: {}", req.getEmail() + " , " + req.getPassword());
 
+        // 이메일 중복 검사
         if (userMapper.findByEmail(req.getEmail()) != null) {
-            throw new BusinessException(ErrorCode.USER_ALREADY_EXISTS);
+            throw new EmailAlreadyVerifiedException();
         }
 
+        // 인증 완료 플래그 확인(필수)
         if (!isEmailVerifiedNow(req.getEmail())) {
-            throw new BusinessException(ErrorCode.EMAIL_NOT_VERIFIED);
+            throw new EmailNotVerifiedException();
         }
 
-        User user = req.toUser();
-        user.setPassword(encoder.encode(user.getPassword()));
-        user.setIsVerified(true);
-        user.setIsActive(true);
-        user.setCreatedAt(LocalDateTime.now());
-        user.setUpdatedAt(LocalDateTime.now());
-        user.setLastPwChangeAt(LocalDateTime.now());
-        user.setRole(org.scoula.user.enums.UserRole.USER);
+        User user = req.toUser(); // DTO → 도메인 객체
+        user.setPassword(encoder.encode(user.getPassword())); // 비밀번호 암호화
 
+        // 기본값 설정
+        user.setIsVerified(true); // 가입 시 true로 저장
+        user.setCreatedAt(LocalDateTime.now());
+
+        // 저장
         try {
             userMapper.save(user);
         } catch (DuplicateKeyException e) {
-            throw new BusinessException(ErrorCode.USER_ALREADY_EXISTS);
+            throw new EmailAlreadyVerifiedException();
         }
 
+        // 가입 완료 후 ok 플래그 제거
         redisService.delete(redisService.evOkKey(req.getEmail()));
 
         return UserResponseDTO.builder()
-                .id(user.getId())
+                .id(user.getUserId())
                 .email(user.getEmail())
-                .userName(user.getUserName())
+                .userName(user.getName())
                 .createdAt(user.getCreatedAt().toString())
-                .nickname(req.getNickname()) // 수정된 부분
                 .build();
     }
+
 
     public TokenResponseDTO login(UserLoginRequestDTO req) {
         log.info("🔒 로그인 시도: {}", req.getEmail());
         User u = userMapper.findByEmail(req.getEmail());
 
         if (u == null) {
-            throw new BusinessException(ErrorCode.USER_NOT_FOUND);
+            throw new ValidationFailedException();
         }
 
         if (!encoder.matches(req.getPassword(), u.getPassword())) {
-            throw new BusinessException(ErrorCode.INVALID_PASSWORD);
+            throw new InvalidPasswordException();
         }
 
-        if (!u.getIsActive()) {
-            throw new BusinessException(ErrorCode.USER_ACCOUNT_DISABLED);
+        // 계정이 활성화 상태인지 확인합니다.
+        if (u.getDeletedAt() != null) {
+            throw new DisabledException("비활성화된 계정입니다.");
         }
 
-        String at = jwtUtil.generateAccessToken(u.getId(), u.getEmail());
-        String rt = jwtUtil.generateRefreshToken(u.getId(), u.getEmail());
-        redisService.saveRefreshToken(u.getId(), rt);
+
+        String at = jwtUtil.generateAccessToken(u.getUserId(), u.getEmail());
+        String rt = jwtUtil.generateRefreshToken(u.getUserId(), u.getEmail());
+        redisService.saveRefreshToken(u.getUserId(), rt);
         return new TokenResponseDTO(at, rt);
     }
 
+
     public TokenResponseDTO refresh(String refreshToken) {
         log.info("🔒 토큰 재발급 시도: {}", refreshToken);
-        if (!jwtUtil.validateToken(refreshToken)) {
-            throw new BusinessException(ErrorCode.JWT_TOKEN_INVALID);
-        }
+        if (!jwtUtil.validateToken(refreshToken))
+            throw new InvalidTokenException();
 
         Long id = jwtUtil.getIdFromToken(refreshToken);
         String email = jwtUtil.getEmailFromToken(refreshToken);
         String saved = redisService.getRefreshToken(id);
-        if (!refreshToken.equals(saved)) {
-            throw new BusinessException(ErrorCode.JWT_TOKEN_INVALID); // 재사용된 리프레시 토큰은 유효하지 않은 것으로 간주
-        }
+        if (!refreshToken.equals(saved))
+            throw new ExpiredTokenException();
 
         String at = jwtUtil.generateAccessToken(id, email);
         String rt = jwtUtil.generateRefreshToken(id, email);
@@ -172,45 +206,54 @@ public class UserServiceImpl implements UserService {
         return new TokenResponseDTO(at, rt);
     }
 
-    public void logout(String bearerToken) {
+    // 로그아웃 시, refreshToken 삭제 및 블랙리스트에 accessToken 추가
+    public void logout(String bearerToken){
+
         if (bearerToken == null || !bearerToken.startsWith("Bearer ")) {
             log.warn("로그아웃 요청 실패: 유효하지 않은 토큰 형식");
-            throw new BusinessException(ErrorCode.JWT_TOKEN_INVALID);
+            throw new InvalidTokenException();
         }
         log.info("로그아웃 시도: {}", bearerToken);
 
-        String token = bearerToken.substring("Bearer ".length());
-        Long userId = jwtUtil.getIdFromToken(token);
+        //1. 토큰 검증
+        if (bearerToken.startsWith("Bearer ")) {
 
-        redisService.deleteRefreshToken(userId);
-        redisService.blacklistAccessToken(token);
-        log.info("로그아웃 성공: 사용자 ID {}", userId);
+            String token = bearerToken.substring("Bearer ".length());
+
+            // 2. 토큰에서 사용자 ID 추출
+            Long userId = jwtUtil.getIdFromToken(token);
+
+            // 3. Redis에서 refreshToken 삭제 및 블랙리스트에 accessToken 추가
+            redisService.deleteRefreshToken(userId);
+            redisService.blacklistAccessToken(token);
+            log.info("로그아웃 성공: {},{}", token, userId);
+        }
     }
 
     public String resetPassword(String email) {
         log.info("🔒 비밀번호 재발급 시도: {}", email);
         User u = userMapper.findByEmail(email);
-        if (u == null) {
-            throw new BusinessException(ErrorCode.USER_NOT_FOUND);
-        }
-        String temp = UUID.randomUUID().toString().substring(0, 8);
-        u.setPassword(encoder.encode(temp));
-        userMapper.updatePassword(u);
-        return temp;
+        if (u == null) throw new UserNotFoundException();
+        String temp = UUID.randomUUID().toString().substring(0,8); // 임시 비밀번호 생성
+        u.setPassword(encoder.encode(temp)); // 암호화
+        userMapper.updatePassword(u); // DB에 저장
+        return temp;  // TODO: 이메일 발송으로 대체
     }
 
+
     public void withdrawal(String bearerToken) {
+
         String token = bearerToken.substring("Bearer ".length());
 
-        if (!jwtUtil.validateToken(token)) {
-            throw new BusinessException(ErrorCode.JWT_TOKEN_INVALID);
-        }
+        if (!jwtUtil.validateToken(token))
+            throw new InvalidTokenException();
 
         Long id = jwtUtil.getIdFromToken(token);
-        log.info("회원 탈퇴 시도: 사용자 ID {}", id);
+        log.info("회원 탈퇴 시도: {}",id);
 
         userMapper.updateIsActive(id);
         redisService.deleteRefreshToken(id);
         redisService.blacklistAccessToken(token);
     }
+
 }
